:sparkles: refactor player and character classes; add caching to DBSC methods; enhance StringUtils with table dumping and string generation

+++ src/main/Content/Script/xgen-core/server/util/dbsc.lua
@@ -1,5 +1,6 @@
 ---@class DBSC (Database Synchronized Class)
 ---@field private __meta DBSC.Meta
+---@field private __cache table<string, DBSC> cached objects
 DBSC = {}
 DBSC.__index = DBSC
 
@@ -20,6 +21,7 @@ DBSC.__index = DBSC
 function DBSC:new(meta)
     local instance = setmetatable({}, self)
     instance.__meta = meta
+    instance.__cache = {}
     return instance
 end
 
@@ -37,8 +39,11 @@ function DBSC:init()
         if key.foreign_key then
             sql = sql .. " REFERENCES " .. key.foreign_key
         end
+        if key.default then
+            sql = sql .. " DEFAULT " .. key.default
+        end
         if i < #self.__meta.columns then
-            sql = sql .. ", "
+            sql = sql .. " NOT NULL, "
         end
     end
     sql = sql .. ")"
@@ -50,6 +55,8 @@ end
 ---@nodiscard
 ---@return boolean success whether the insert was successful
 function DBSC:insert()
+    local pks = {}
+    local pksString = ""
     local sql = "INSERT INTO " .. self.__meta.name .. " ("
     local values = {}
 
@@ -57,6 +64,16 @@ function DBSC:insert()
 
     for i, key in ipairs(self.__meta.columns) do
         if key.auto_increment then
+            goto continue
+        end
+        if key.primary_key then
+            pks[key.name] = self[key.name]
+            pksString = pksString .. ":" .. tostring(self[key.name])
+        end
+        if key.default then
+            if i >= #self.__meta.columns then
+                sql = sql:sub(1, -3)
+            end
             goto continue
         end
         sql = sql .. key.name
@@ -69,7 +86,13 @@ function DBSC:insert()
     end
     sql = sql .. ") VALUES (" .. table.concat(values, ", ") .. ")"
 
-    return Database.Execute(sql, thisVals)
+    if not Database.Execute(sql, thisVals) then
+        return false
+    end
+
+    self.__cache[pksString] = self:get(pks)
+
+    return true
 end
 
 
@@ -106,6 +129,15 @@ end
 ---@param primary_keys table<string, any> the values of the primary keys
 ---@return DBSC? object The loaded object or nil if not found
 function DBSC:get(primary_keys)
+    local pksString = ""
+    for _, value in pairs(primary_keys) do
+        pksString = pksString .. ":" .. tostring(value)
+    end
+
+    if self.__cache[pksString] then
+        return self.__cache[pksString]
+    end
+
     local sql = "SELECT * FROM " .. self.__meta.name .. " WHERE "
     local values = {}
     local conditions = {}
@@ -127,16 +159,59 @@ function DBSC:get(primary_keys)
         for _, key in ipairs(self.__meta.columns) do
             instance[key.name] = result[string.upper(key.name)]
         end
+
+        self.__cache[pksString] = instance
+
         return instance
     end
 
     return nil
 end
 
+function DBSC:getWhere(conditions)
+    local sql = "SELECT * FROM " .. self.__meta.name .. " WHERE "
+    local values = {}
+    local conds = {}
+
+    for key, value in pairs(conditions) do
+        table.insert(conds, key .. " = ?")
+        table.insert(values, value)
+    end
+
+    sql = sql .. table.concat(conds, " AND ")
+
+    local results = Database.Select(sql, values)
+    if not results then return nil end
+
+    local instances = {}
+    for _, result in ipairs(results) do
+        local instance = {}
+        setmetatable(instance, self)
+        for _, key in ipairs(self.__meta.columns) do
+            instance[key.name] = result.Column[string.upper(key.name)]
+        end
+        table.insert(instances, instance)
+
+    end
+
+    return instances
+end
+
 ---Deletes the object from the database.
 ---@nodiscard
 ---@return boolean success whether the delete was successful
 function DBSC:delete()
+    local pksString = ""
+    for _, key in ipairs(self.__meta.columns) do
+        if key.primary_key then
+            pksString = pksString .. ":" .. tostring(self[key.name])
+        end
+    end
+
+    if not self.__cache[pksString] then
+        return true
+    end
+
     local sql = "DELETE FROM " .. self.__meta.name .. " WHERE "
     local values = {}
     local conditions = {}
@@ -149,5 +224,11 @@ function DBSC:delete()
     end
 
     sql = sql .. table.concat(conditions, " AND ")
-    return Database.Execute(sql, values)
+    if not Database.Execute(sql, values) then
+        return false
+    end
+
+    self.__cache[pksString] = nil
+
+    return true
 end
